Complete 2021 day 15 challenge 2
The A* algorithm I wrote initially still worked fine! Most of this
one was generating the expanded map. I could have made it generate
the tile information on the fly, however just pregenerating the map
was less thinking.

If I had a generic A* implementation I was using, it probably would've
been easier to generate the values on the fly.