Complete 2022 day 19 challenge 1
Woo-hoo! This took me forever. It required culling massive parts of the
search space to run efficiently, and I had no idea how to do that.

I don't think this is the most elegant solution, but I'm glad to be
getting some practice on stuff I don't know very well.